requirements.txt
argparse==1.4.*
requests==2.27.*
environs==9.5.*
python_telegram_bot==13.11
